requirements.txt
python-telegram-bot[job-queue]==20.2
requests
Pillow